Final Build
Created a start menu, and a game completion menu. Hooked up the main level to both of these menus, and exported the completed game. It is very short and simple. Some features that are complete were not implemented in levels, and I may do so at a later date if I feel like it. I have decide to leave the game as is to move on to other more interesting ideas.